chore: refactoring into more component-based stucturing of locations of user interface elements
