+++ src/components/header/HeaderComponent.tsx
@@ -1,47 +1,9 @@
 import { handleStepChange } from "@/utilities/setUpUtilities";
-import { handleToggleAllRepos } from "@/utilities/repoDetailUtilities";
-import ButtonCustom from "./ButtonCustom";
-import CheckBoxCustom from "./CheckBoxCustom";
-import "../App.css";
-import {
-  HeaderProps,
-  TitleProps,
-  SelectAllButtonProps,
-} from "@/models/HeaderComponentModels.ts";
-import { Button } from "./ui/button";
-
-const TitleComponent = ({ currentStep, hasPAT }: TitleProps): JSX.Element => {
-  let stepTitle = "";
-
-  switch (currentStep) {
-    case 1:
-      stepTitle = "Choose a Setup Option from Below";
-      break;
-    case 2:
-      if (hasPAT !== null) {
-        stepTitle = "Enter Your PAT and the Github Username/Org Name to Track";
-      } else {
-        stepTitle = "Enter the Github Username/Org Name to Track";
-      }
-      break;
-    case 3:
-      stepTitle = "Choose Repositories";
-      break;
-    case 4:
-      stepTitle = "Your Repositories";
-      break;
-  }
-
-  return (
-    <div style={{ display: "flex", justifyContent: "center" }}>
-      <div>
-        <h2 className="stepTitle" style={{ textAlign: "left" }}>
-          {stepTitle}
-        </h2>
-      </div>
-    </div>
-  );
-};
+import ButtonCustom from "../input/ButtonCustom";
+import "../../App.css";
+import { HeaderProps } from "@/models/HeaderComponentModels.ts";
+import SelectAllButton from "../input/SelectAllButton";
+import TitleComponent from "./TitleComponent";
 
 const HeaderComponent = ({
   setStepState,
@@ -148,91 +110,6 @@ const HeaderComponent = ({
           setRepoDetails={setRepoDetails}
         />
       )}
-    </div>
-  );
-};
-
-const SelectAllButton = ({
-  allReposToggled,
-  repoDetails,
-  activeNumPRs,
-  setActiveNumPRs,
-  setRepoDetails,
-  setReposToggled,
-}: SelectAllButtonProps): JSX.Element => {
-  async function toggleAllSelectedRepos() {
-    if (repoDetails !== null) {
-      setReposToggled(!allReposToggled);
-      await handleToggleAllRepos({
-        allReposToggled,
-        setRepoDetails,
-        setActiveNumPRs,
-        activeNumPRs,
-        repoDetails,
-      });
-    }
-  }
-  console.log(
-    "Select all button will receive repoChecked value: ",
-    allReposToggled
-  );
-
-  return (
-    <div
-      style={{
-        display: "flex",
-        justifyContent: "flex-end",
-        marginTop: "5px",
-        marginBottom: "5px",
-        marginRight: "10px",
-      }}
-    >
-      <Button
-        variant="secondary"
-        className="rounded shadow mb-5"
-        id="select-all-button"
-        onClick={toggleAllSelectedRepos}
-        style={{
-          paddingBottom: "15px",
-          paddingTop: "15px",
-          paddingRight: "5px",
-        }}
-      >
-        <div
-          style={{
-            display: "flex",
-            justifyContent: "center",
-            flexDirection: "row",
-            width: "100%",
-          }}
-        >
-          <div
-            style={{
-              display: "flex",
-              justifyContent: "center",
-              flexDirection: "column",
-            }}
-          >
-            <p className="adjacentButtonText">Select All</p>
-          </div>
-          <div
-            style={{
-              display: "flex",
-              justifyContent: "center",
-              flexDirection: "column",
-              marginTop: "2px",
-              marginLeft: "10px",
-              marginRight: "0",
-            }}
-          >
-            <CheckBoxCustom
-              handleClick={toggleAllSelectedRepos}
-              repoChecked={allReposToggled}
-              name={"-ToggleAll"}
-            />
-          </div>
-        </div>
-      </Button>
     </div>
   );
 };
