+++ src/models/HeaderComponentModels.ts
@@ -26,15 +26,4 @@ type TitleProps = {
   hasPAT: string | null;
 };
 
-type SelectAllButtonProps = {
-  allReposToggled: boolean;
-  repoDetails: RepoCardComponentDetails[] | null;
-  activeNumPRs: ActiveNumPRs[];
-  setReposToggled: React.Dispatch<React.SetStateAction<boolean>>;
-  setActiveNumPRs: React.Dispatch<React.SetStateAction<ActiveNumPRs[]>>;
-  setRepoDetails: React.Dispatch<
-    React.SetStateAction<RepoCardComponentDetails[] | null>
-  >;
-};
-
-export type { HeaderProps, TitleProps, SelectAllButtonProps };
+export type { HeaderProps, TitleProps };
